+++ Framework/AmplitudeFramework.h
@@ -4,3 +4,10 @@
 #import <Amplitude/AMPRevenue.h>
 #import <Amplitude/AMPTrackingOptions.h>
 #import <Amplitude/Amplitude+SSLPinning.h>
+#import <Amplitude/AMPPlan.h>
+#import <Amplitude/AMPMiddleware.h>
+#import <Amplitude/AMPServerZone.h>
+
+#if TARGET_OS_WATCH
+#import <Amplitude/AMPBackgroundNotifier.h>
+#endif