+++ Tests/AmplitudeTests.m
@@ -15,6 +15,8 @@
 #import "AMPDeviceInfo.h"
 #import "AMPUtils.h"
 #import "AMPTrackingOptions.h"
+#import "AMPPlan.h"
+#import "AMPServerZone.h"
 
 @interface Amplitude (Tests)
 
@@ -1084,6 +1086,75 @@
     
     NSDictionary *currentLibraryValue = event[@"library"];
     XCTAssertEqualObjects(currentLibraryValue, targetLibraryValue);
+}
+
+- (void)testSetPlan {
+    Amplitude *client = [Amplitude instanceWithName:@"observe_plan"];
+    [client initializeApiKey:@"tracking_plan"];
+    NSString *branch = @"main";
+    NSString *source = @"mobile";
+    NSString *version = @"1.0.0";
+    
+    AMPPlan *plan = [[[[AMPPlan plan] setBranch:branch] setSource:source] setVersion:version];
+    [client setPlan:plan];
+    [client logEvent:@"test"];
+    [client flushQueue];
+    NSDictionary *event = [client getLastEventFromInstanceName:@"observe_plan"];
+    
+    NSDictionary *planValue = event[@"plan"];
+    XCTAssertEqualObjects(branch, planValue[@"branch"]);
+    XCTAssertEqualObjects(source, planValue[@"source"]);
+    XCTAssertEqualObjects(version, planValue[@"version"]);
+}
+
+- (void)testSetServerZone {
+    Amplitude *client = [Amplitude instanceWithName:@"eu_zone"];
+    XCTAssertEqualObjects(kAMPEventLogUrl, [client valueForKey:@"serverUrl"]);
+    [client initializeApiKey:@"eu_api_key"];
+    [client setServerZone:EU];
+    XCTAssertEqualObjects(kAMPEventLogEuUrl, [client valueForKey:@"serverUrl"]);
+}
+
+- (void)testSetServerZoneWithoutUpdateServerUrl {
+    Amplitude *client = [Amplitude instanceWithName:@"eu_zone_2"];
+    XCTAssertEqualObjects(kAMPEventLogUrl, [client valueForKey:@"serverUrl"]);
+    [client initializeApiKey:@"eu_api_key"];
+    [client setServerZone:EU updateServerUrl:NO];
+    XCTAssertEqualObjects(kAMPEventLogUrl, [client valueForKey:@"serverUrl"]);
+}
+
+- (void)testMiddlewareSupport {
+    NSString *eventType = @"middleware event";
+    AMPBlockMiddleware *updateEventTypeMiddleware = [[AMPBlockMiddleware alloc] initWithBlock: ^(AMPMiddlewarePayload * _Nonnull payload, AMPMiddlewareNext _Nonnull next) {
+        [payload.event setValue:eventType forKey:@"event_type"];
+        [payload.event setValue:payload.extra[@"description"] forKey:@"description"];
+        next(payload);
+    }];
+    Amplitude *client = [Amplitude instanceWithName:@"middleware_support"];
+    [client addEventMiddleware:updateEventTypeMiddleware];
+    [client initializeApiKey:@"middleware_api_key"];
+    NSMutableDictionary *eventProperties = [NSMutableDictionary dictionary];
+    [eventProperties setObject:@"green" forKey:@"color"];
+    NSMutableDictionary *middlewareExtra = [NSMutableDictionary dictionary];
+    [middlewareExtra setObject:@"some event description" forKey:@"description"];
+    [client logEvent:@"test" withEventProperties:eventProperties withMiddlewareExtra:middlewareExtra];
+    [client flushQueue];
+
+    NSDictionary *event = [client getLastEventFromInstanceName:@"middleware_support"];
+    XCTAssertEqualObjects(eventType, event[@"event_type"]);
+    XCTAssertEqualObjects(middlewareExtra[@"description"], event[@"description"]);
+}
+
+- (void)testSwallowMiddleware {
+    AMPBlockMiddleware *swallowMiddleware = [[AMPBlockMiddleware alloc] initWithBlock: ^(AMPMiddlewarePayload * _Nonnull payload, AMPMiddlewareNext _Nonnull next) {
+    }];
+    Amplitude *client = [Amplitude instanceWithName:@"middleware_swallow"];
+    [client addEventMiddleware:swallowMiddleware];
+    [client initializeApiKey:@"middleware_api_key"];
+    [client logEvent:@"test"];
+    [client flushQueue];
+    
+    XCTAssertNil([client getLastEventFromInstanceName:@"middleware_swallow"]);
 }
 
 @end